+++ vectortypes.h
@@ -12,7 +12,7 @@ namespace anl
 		SRGBA() : r(0), g(0), b(0), a(0){}
 		SRGBA(int al) : r(a), g(a), b(a), a(al) {}
 		SRGBA(const  SRGBA &rhs) : r(rhs.r), g(rhs.g), b(rhs.b), a(rhs.a){}
-		SRGBA(const float t1, const float t2, const float t3, const float t4) : r(t1), g(t2), b(t3), a(t4){}
+		SRGBA(const double t1, const double t2, const double t3, const double t4) : r(t1), g(t2), b(t3), a(t4){}
 		~SRGBA(){}
 
     bool operator ==(const SRGBA &rhs) const
@@ -156,10 +156,10 @@ namespace anl
         return (*this);
     };
 
-		float r;
-		float g;
-		float b;
-		float a;
+		double r;
+		double g;
+		double b;
+		double a;
 	};
 };
